refactor: rename folder restclient
