+++ src/components/restClient/CodeSpace.tsx
@@ -4,7 +4,7 @@ import { Editor } from '@monaco-editor/react';
 import { Card, Flex, Select, Typography } from 'antd';
 import { useState } from 'react';
 
-import type { ResponseInfo } from '@/components/rest-client/types/rest-client';
+import type { ResponseInfo } from '@/components/restClient/types/rest-client';
 
 type Props = {
   value: string;
